+++ src/luzombra/geometry/linearAlgebra/TransMat.scala
@@ -6,7 +6,7 @@ package luzombra.geometry.linearAlgebra
   * @param y translation along y-axis
   * @param z translation along z-axis
   */
-class TransMat(val x: Double, val y: Double, val z: Double) extends Mat4x4(
+class TransMat(val x: Float, val y: Float, val z: Float) extends Mat4x4(
   (1, 0, 0, 0), (0, 1, 0, 0), (0, 0, 1, 0), (x, y, z, 1)) {
 
   /** Returns a transMat that reverses the translation by this transMat */
@@ -17,6 +17,6 @@ class TransMat(val x: Double, val y: Double, val z: Double) extends Mat4x4(
 object TransMat {
 
   /** Returns a new transMat */
-  def apply(x: Double, y: Double, z: Double): TransMat = new TransMat(x, y, z)
+  def apply(x: Float, y: Float, z: Float): TransMat = new TransMat(x, y, z)
 
 }
